blind code. Did not test

+++ util.h
@@ -1,2 +1,2 @@
-char *binary_to_hex(void *data, ssize_t n);
-void *hex_to_binary(char *hex, ssize_t *bin_bytes);
+char *binary_to_hex(void *data, int  n);
+void *hex_to_binary(char *hex, int *bin_bytes);
